+++ CHANGELOG.md
@@ -77,6 +77,7 @@
 - PR #2326 Fix for bug in DLPack when reading multiple columns
 - PR #2324 Fix cudf Docker build
 - PR #2325 Fix ORC RLEv2 patched base mode decoding with nonzero patch width
+- PR #2235 Fix get_dummies to be compatible with dask
 
 
 # cuDF 0.8.0 (27 June 2019)